:see_no_evil: Removed amplifyconfiguration.dart form gitigniore

+++ libs/redux_comp/lib/models/chat/message_model.dart
@@ -17,6 +17,6 @@ class MessageModel {
   factory MessageModel.fromJson(Map<String, dynamic> json) => MessageModel(
         msg: json['msg'],
         sender: json['sender'],
-        timestamp: json['timestamp'],
+        timestamp: json['timestamp'].toDouble(),
       );
 }
